reformat the code
